'nohidden': on_bufclosed: non-aggressive prev-buf
If user has 'nohidden' and Dirvish is being used to open the first
non-directory buffer, then forcing visit_prevbuf() causes a :noautoload
visit to the target buffer, which means it is loaded without syntax
highlighting.

Note: there's still another case: if user loads Dirvish with 'nohidden',
the prevbuf is _unloaded_ (not hidden). When Dirvish returns to it,
again it did so with :noautolad which means syntax is missing.

References #13