Do not set 'nomodifiable'.
This was a holdover from common practices in other plugins. But it's
not needed here because Dirvish never modifies the filesystem, so the
risk of accidents is minimal. Allowing 'modifiable' also makes
customization and ad-hoc :g/:s/:sort more convenient.

We still nuke the local undo tree on setup (and reload), to save memory.