buf_render: Skip housekeeping if buffer is new.
On reload we try very hard to preserve the viewport of all windows
showing the same directory. But if the buffer is empty (i.e., new) then
that isn't needed. To check "empty" we only test the first line, but
again this is low-impact, user requested a reload.

This fixes an edge case with:
    vim -p ...
    vim -o ...
    vim -O ...
where the buffer is added to multiple windows/tabpages before
BufEnter/BufAdd is fired.

Fixes #30