Use shellslash to determine path separator
shellslash should tell vim what slash to pass to the shell. Following
that logic, the code shouldn't check 'shell' at all. Instead, it should
only use shellslash:

    exists('+shellslash') && &shellslash == 0

I doubt cmd.exe would be used as a shell on platforms that don't have
shellslash. shellslash only exists on Windows and evaluates to 0 on
Linux (tested gvim 7.4.52 on Ubuntu). So this change shouldn't break
other platforms.

Now dirvish uses conceal to hide the directories in the listing when
shellslash is enabled with cmd.exe (on Windows).

Tested on gvim 7.4.1832 on Win8.1 with 'shellslash' and 'noshellslash'.
dirvish correctly conceals paths.