fix(apply_icons): filename may be truncated #243
Problem:
dirvish#add_icon_fn() may truncate filenames from the left, when icons
are being applied.

To see the issue, create files 'abc-file.txt', 'defg-file.txt', and
'file.txt' in a directory and display them in a dirvish buffer.  All
filenames display as 'file.txt'.

The problem seems to occur when the basename of one of the files in the
directory matches the right-hand portion of one or more other filenames
in the same directory.  However, experimentation shows that it does not
always happen in this case.

Solution:
Define each pattern to match an exact line.